requirements.txt
beautifulsoup4==4.12.2
chromadb==0.3.26
fastapi==0.97.0
gradio==3.35.2
langchain==0.0.208
pydantic==1.10.9
python-dotenv==1.0.0
uvicorn==0.22.0
